Renames a bunch of components and moves HTML template from transaction-children-list-table-base to transaction-guarantors
