+++ front-end/src/app/reports/transactions/double-transaction-detail/double-transaction-detail.component.html
@@ -68,7 +68,7 @@
   </p-accordionTab>
 </p-accordion>
 
-<app-transaction-children-tables [transaction]="transaction"></app-transaction-children-tables>
+<app-transaction-children-list-container [transaction]="transaction"></app-transaction-children-list-container>
 
 <app-transaction-navigation
   [isEditable]="isEditable"
