+++ front-end/src/app/reports/transactions/transaction-children-list-container/transaction-children-list-container.component.ts
@@ -2,10 +2,10 @@ import { Component, Input } from '@angular/core';
 import { Transaction } from 'app/shared/models/transaction.model';
 
 @Component({
-  selector: 'app-transaction-children-tables',
-  templateUrl: 'transaction-children-tables.component.html',
+  selector: 'app-transaction-children-list-container',
+  templateUrl: 'transaction-children-list-container.component.html',
   styleUrls: ['../transaction.scss'],
 })
-export class TransactionChildrenTablesComponent {
+export class TransactionChildrenListContainerComponent {
   @Input() transaction?: Transaction;
 }
