+++ front-end/src/app/reports/transactions/transaction-detail/transaction-detail.component.html
@@ -21,7 +21,7 @@
   </form>
 </div>
 
-<app-transaction-children-tables [transaction]="transaction"></app-transaction-children-tables>
+<app-transaction-children-list-container [transaction]="transaction"></app-transaction-children-list-container>
 
 <app-transaction-navigation
   [isEditable]="isEditable"
