+++ front-end/src/app/reports/transactions/transaction-list/transaction-guarantors/transaction-guarantors.component.html
@@ -52,7 +52,9 @@
           </a>
         </td>
         <td>
-          {{ item[item.transactionType.templateMap.amount] | currency }}
+          <span class="float-right">
+            {{ item[item.transactionType.templateMap.amount] | currency }}
+          </span>
         </td>
         <td>
           <app-table-actions-button
