+++ front-end/src/app/reports/transactions/transaction.scss
@@ -43,3 +43,7 @@
   display: flex;
   justify-content: space-between;
 }
+
+.float-right {
+  float: right;
+}
