+++ front-end/src/app/reports/transactions/triple-transaction-detail/triple-transaction-detail.component.html
@@ -102,7 +102,7 @@
   </p-accordionTab>
 </p-accordion>
 
-<app-transaction-children-tables [transaction]="transaction"></app-transaction-children-tables>
+<app-transaction-children-list-container [transaction]="transaction"></app-transaction-children-list-container>
 
 <app-transaction-navigation
   [isEditable]="isEditable"
